--- app/src/main/java/com/intellistart/interviewplanning/slot/application/port/NatsEventPublisherOutPort.kt
@@ -1,8 +1,0 @@
-package com.intellistart.interviewplanning.slot.application.port
-
-import com.intellistart.interviewplanning.output.pubsub.slot.SlotUpdatedEvent
-
-interface NatsEventPublisherOutPort {
-
-    fun publish(slotUpdatedEvent: SlotUpdatedEvent)
-}
