--- app/src/main/java/com/intellistart/interviewplanning/slot/application/port/SlotKafkaProducerOutPort.kt
@@ -1,8 +1,0 @@
-package com.intellistart.interviewplanning.slot.application.port
-
-import com.intellistart.interviewplanning.slot.domain.model.Slot
-
-interface SlotKafkaProducerOutPort {
-
-    fun produceNotification(slot: Slot, id: String)
-}
