+++ app/src/main/java/com/intellistart/interviewplanning/slot/infrastructure/adapters/kafka/producer/Producer.kt
@@ -2,7 +2,7 @@ package com.intellistart.interviewplanning.slot.infrastructure.adapters.kafka.pr
 
 import com.intellistart.interviewplanning.KafkaTopic
 import com.intellistart.interviewplanning.output.pubsub.slot.SlotUpdatedEvent
-import com.intellistart.interviewplanning.slot.application.port.SlotKafkaProducerOutPort
+import com.intellistart.interviewplanning.slot.port.ProducerOutPort
 import com.intellistart.interviewplanning.slot.domain.model.Slot
 import com.intellistart.interviewplanning.slot.infrastructure.dto.toProto
 import org.apache.kafka.clients.producer.ProducerRecord
@@ -12,11 +12,11 @@ import reactor.kafka.sender.SenderRecord
 import reactor.kotlin.core.publisher.toMono
 
 @Component
-class SlotKafkaProducer(
+class Producer(
     private val kafkaSender: KafkaSender<String, SlotUpdatedEvent>
-) : SlotKafkaProducerOutPort {
+) : ProducerOutPort {
 
-    override fun produceNotification(slot: Slot, id: String) {
+    override fun produceSlotNotificationToKafka(slot: Slot, id: String) {
         val event = SlotUpdatedEvent.newBuilder().setSlotId(id).setSlot(slot.toProto()).build()
         kafkaSender.send(
             SenderRecord.create(
