+++ app/src/main/java/com/intellistart/interviewplanning/slot/infrastructure/adapters/nats/event/EventPublisher.kt
@@ -2,14 +2,14 @@ package com.intellistart.interviewplanning.slot.infrastructure.adapters.nats.eve
 
 import com.intellistart.interviewplanning.NatsSubject
 import com.intellistart.interviewplanning.output.pubsub.slot.SlotUpdatedEvent
-import com.intellistart.interviewplanning.slot.application.port.NatsEventPublisherOutPort
+import com.intellistart.interviewplanning.slot.port.EventPublisherOutPort
 import io.nats.client.Connection
 import org.springframework.stereotype.Component
 
 @Component
-class NatsEventPublisher(private val connection: Connection) : NatsEventPublisherOutPort {
+class EventPublisher(private val connection: Connection) : EventPublisherOutPort {
 
-    override fun publish(slotUpdatedEvent: SlotUpdatedEvent) {
+    override fun publishNatsSlotUpdatedEvent(slotUpdatedEvent: SlotUpdatedEvent) {
         val updateEventSubject = NatsSubject.createSlotEventNatsSubject(slotUpdatedEvent.slotId, "UPDATE")
         connection.publish(updateEventSubject, slotUpdatedEvent.toByteArray())
     }
