+++ app/src/main/java/com/intellistart/interviewplanning/slot/infrastructure/adapters/nats/slot/GetSlotsNatsController.kt
@@ -4,7 +4,7 @@ import com.google.protobuf.Parser
 import com.intellistart.interviewplanning.NatsSubject
 import com.intellistart.interviewplanning.request.slot.get_all.proto.GetAllSlotsRequest
 import com.intellistart.interviewplanning.request.slot.get_all.proto.GetAllSlotsResponse
-import com.intellistart.interviewplanning.slot.application.port.SlotOperationsInPort
+import com.intellistart.interviewplanning.slot.port.SlotOperationsInPort
 import com.intellistart.interviewplanning.slot.domain.model.Slot
 import com.intellistart.interviewplanning.slot.infrastructure.adapters.nats.NatsController
 import com.intellistart.interviewplanning.slot.infrastructure.dto.toProto
