+++ app/src/main/java/com/intellistart/interviewplanning/slot/infrastructure/mapper/SlotMapper.kt
@@ -3,7 +3,7 @@ package com.intellistart.interviewplanning.slot.infrastructure.mapper
 import com.intellistart.interviewplanning.infrastructure.adapters.reopository.entity.toDomain
 import com.intellistart.interviewplanning.infrastructure.adapters.reopository.entity.toEntity
 import com.intellistart.interviewplanning.slot.domain.model.Slot
-import com.intellistart.interviewplanning.slot.infrastructure.adapters.reopository.entity.SlotEntity
+import com.intellistart.interviewplanning.slot.infrastructure.adapters.reopository.mongo.entity.SlotEntity
 import org.bson.types.ObjectId
 
 fun Slot.toEntity(): SlotEntity {
