+++ app/src/main/java/com/intellistart/interviewplanning/slot/port/EventPublisherOutPort.kt
@@ -1,0 +1,8 @@
+package com.intellistart.interviewplanning.slot.port
+
+import com.intellistart.interviewplanning.output.pubsub.slot.SlotUpdatedEvent
+
+interface EventPublisherOutPort {
+
+    fun publishNatsSlotUpdatedEvent(slotUpdatedEvent: SlotUpdatedEvent)
+}
