+++ app/src/main/java/com/intellistart/interviewplanning/slot/port/EventSubscriberInPort.kt
@@ -1,0 +1,9 @@
+package com.intellistart.interviewplanning.slot.port
+
+import com.intellistart.interviewplanning.output.pubsub.slot.SlotUpdatedEvent
+import reactor.core.publisher.Flux
+
+interface EventSubscriberInPort {
+
+    fun subscribeOnNatsSlotUpdatedEvent(slotId: String): Flux<SlotUpdatedEvent>
+}
