+++ app/src/main/java/com/intellistart/interviewplanning/slot/port/ProducerOutPort.kt
@@ -1,0 +1,8 @@
+package com.intellistart.interviewplanning.slot.port
+
+import com.intellistart.interviewplanning.slot.domain.model.Slot
+
+interface ProducerOutPort {
+
+    fun produceSlotNotificationToKafka(slot: Slot, id: String)
+}
